Fix a big race-condition bug
I was relying on createNode to return a promise which resolves when the
onCreateNode hook completes for the newly created node. It turns out
that I was mistaken that this was the way it worked. I reworked my
onCreateNode hook to listen for both MarkdownRemark nodes created
originally and created as a result of my plugin, and used some shared
state to determine when and how to link the result.